% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descstat.R
\name{descstat}
\alias{descstat}
\title{Generates and displays both graphical summaries and numerical summaries for numeric data}
\usage{
descstat(x)
}
\arguments{
\item{x}{is the vector containing the random sample to be used}
}
\description{
Generates and displays both graphical summaries and numerical summaries for numeric data
}
